First step at refactoring
