+++ commands/run.go
@@ -1,0 +1,669 @@
+package commands
+
+//DockerRun Runs image described by Seed spec
+func DockerRun() {
+	imageName := runCmd.Lookup(constants.ImgNameFlag).Value.String()
+	if imageName == "" {
+		fmt.Fprintf(os.Stderr, "ERROR: No input image specified\n")
+	}
+
+	// Parse seed information off of the label
+	seed := SeedFromImageLabel(imageName)
+
+	// build docker run command
+	dockerArgs := []string{"run"}
+
+	if runCmd.Lookup(constants.RmFlag).Value.String() == constants.TrueString {
+		dockerArgs = append(dockerArgs, "--rm")
+	}
+
+	var mountsArgs []string
+	var envArgs []string
+	var resourceArgs []string
+	var inputSize float64
+	var outputSize float64
+
+	// expand INPUT_FILEs to specified inputData files
+	if seed.Job.Interface.InputData.Files != nil {
+		inMounts, size, temp, err := DefineInputs(&seed)
+		for _, v := range temp {
+			defer RemoveAll(v)
+		}
+		if err != nil {
+			fmt.Fprintf(os.Stderr, "ERROR: Error occurred processing inputData arguments.\n%s", err.Error())
+			fmt.Fprintf(os.Stderr, "Exiting seed...\n")
+			os.Exit(1)
+		} else if inMounts != nil {
+			mountsArgs = append(mountsArgs, inMounts...)
+			inputSize = size
+		}
+	}
+
+	if len(seed.Job.Interface.Resources.Scalar) > 0 {
+		inResources, diskSize, err := DefineResources(&seed, inputSize)
+		if err != nil {
+			fmt.Fprintf(os.Stderr, "ERROR: Error occurred processing resources\n%s", err.Error())
+			fmt.Fprintf(os.Stderr, "Exiting seed...\n")
+			os.Exit(1)
+		} else if inResources != nil {
+			resourceArgs = append(resourceArgs, inResources...)
+			outputSize = diskSize
+		}
+	}
+
+	// mount the JOB_OUTPUT_DIR (outDir flag)
+	var outDir string
+	if strings.Contains(seed.Job.Interface.Cmd, "OUTPUT_DIR") {
+		outDir = SetOutputDir(imageName, &seed)
+		if outDir != "" {
+			mountsArgs = append(mountsArgs, "-v")
+			mountsArgs = append(mountsArgs, outDir+":"+outDir)
+		}
+	}
+
+	// Settings
+	if seed.Job.Interface.Settings != nil {
+		inSettings, err := DefineSettings(&seed)
+		if err != nil {
+			fmt.Fprintf(os.Stderr, "ERROR: Error occurred processing settings arguments.\n%s", err.Error())
+			fmt.Fprintf(os.Stderr, "Exiting seed...\n")
+			os.Exit(1)
+		} else if inSettings != nil {
+			envArgs = append(envArgs, inSettings...)
+		}
+	}
+
+	// Additional Mounts defined in seed.json
+	if seed.Job.Interface.Mounts != nil {
+		inMounts, err := DefineMounts(&seed)
+		if err != nil {
+			fmt.Fprintf(os.Stderr, "ERROR: Error occurred processing mount arguments.\n%s", err.Error())
+			fmt.Fprintf(os.Stderr, "Exiting seed...\n")
+			os.Exit(1)
+		} else if inMounts != nil {
+			mountsArgs = append(mountsArgs, inMounts...)
+		}
+	}
+
+	// Build Docker command arguments:
+	// 		run
+	//		-rm if specified
+	//		env injection
+	// 		all mounts
+	//		image name
+	//		Job.Interface.Cmd
+	dockerArgs = append(dockerArgs, mountsArgs...)
+	dockerArgs = append(dockerArgs, envArgs...)
+	dockerArgs = append(dockerArgs, resourceArgs...)
+	dockerArgs = append(dockerArgs, imageName)
+
+	// Parse out command arguments from seed.Job.Interface.Cmd
+	args := strings.Split(seed.Job.Interface.Cmd, " ")
+	dockerArgs = append(dockerArgs, args...)
+
+	// Run
+	var cmd bytes.Buffer
+	cmd.WriteString("docker ")
+	for _, s := range dockerArgs {
+		cmd.WriteString(s + " ")
+	}
+	fmt.Fprintf(os.Stderr, "INFO: Running Docker command:\n%s\n", cmd.String())
+
+	// Run Docker command and capture output
+	runCmd := exec.Command("docker", dockerArgs...)
+	var errs bytes.Buffer
+	runCmd.Stderr = io.MultiWriter(os.Stderr, &errs)
+	runCmd.Stdout = os.Stderr
+
+	// Run docker run
+	if err := runCmd.Run(); err != nil {
+		fmt.Fprintf(os.Stderr, "ERROR: error executing docker run. %s\n",
+			err.Error())
+	}
+
+	if errs.String() != "" {
+		fmt.Fprintf(os.Stderr, "ERROR: Error running image '%s':\n%s\n",
+			imageName, errs.String())
+		fmt.Fprintf(os.Stderr, "Exiting seed...\n")
+		os.Exit(1)
+	}
+
+	// Validate output against pattern
+	if seed.Job.Interface.OutputData.Files != nil ||
+		seed.Job.Interface.OutputData.JSON != nil {
+		ValidateOutput(&seed, outDir, outputSize)
+	}
+}
+
+//DefineInputs extracts the paths to any input data given by the 'run' command
+// flags 'inputData' and sets the path in the json object. Returns:
+// 	[]string: docker command args for input files in the format:
+//	"-v /path/to/file1:/path/to/file1 -v /path/to/file2:/path/to/file2 etc"
+func DefineInputs(seed *objects.Seed) ([]string, float64, map[string]string, error) {
+
+	// Validate inputs given vs. inputs defined in manifest
+	
+	var mountArgs []string
+	var sizeMiB float64
+
+	// Ingest inputs into a map key = inputkey, value=inputpath
+	inputs := strings.Split(runCmd.Lookup(constants.InputDataFlag).Value.String(), ",")
+	inMap := make(map[string]string)
+	for _, f := range inputs {
+		x := strings.Split(f, "=")
+		if len(x) != 2 {
+			fmt.Fprintf(os.Stderr, "ERROR: Input files should be specified in KEY=VALUE format.\n")
+			fmt.Fprintf(os.Stderr, "ERROR: Unknown key for input %v encountered.\n",
+				x)
+			continue
+		}
+		inMap[x[0]] = x[1]
+	}
+
+	// Valid by default
+	valid := true
+	var keys []string
+	var unrequired []string
+	var tempDirectories map[string]string
+	tempDirectories = make(map[string]string)
+	for _, f := range seed.Job.Interface.InputData.Files {
+		if f.Multiple {
+			tempDir := "temp-" + time.Now().Format(time.RFC3339)
+			tempDir = strings.Replace(tempDir, ":", "_", -1)
+			os.Mkdir(tempDir, os.ModePerm)
+			tempDirectories[f.Name] = tempDir
+			mountArgs = append(mountArgs, "-v")
+			mountArgs = append(mountArgs, GetFullPath(tempDir)+":/"+tempDir)
+		}
+		if f.Required == false {
+			unrequired = append(unrequired, f.Name)
+			continue
+		}
+		keys = append(keys, f.Name)
+		if _, prs := inMap[f.Name]; !prs {
+			valid = false
+		}
+	}
+
+	if !valid {
+		var buffer bytes.Buffer
+		buffer.WriteString("ERROR: Incorrect input data files key/values provided. -i arguments should be in the form:\n")
+		buffer.WriteString("  seed run -i KEY1=path/to/file1 -i KEY2=path/to/file2 ...\n")
+		buffer.WriteString("The following input file keys are expected:\n")
+		for _, n := range keys {
+			buffer.WriteString("  " + n + "\n")
+		}
+		buffer.WriteString("\n")
+		return nil, 0.0, tempDirectories, errors.New(buffer.String())
+	}
+
+	for _, f := range inputs {
+		x := strings.Split(f, "=")
+		if len(x) != 2 {
+			fmt.Fprintf(os.Stderr, "ERROR: Input files should be specified in KEY=VALUE format.\n")
+			fmt.Fprintf(os.Stderr, "ERROR: Unknown key for input %v encountered.\n",
+				inputs)
+			continue
+		}
+
+		key := x[0]
+		val := x[1]
+
+		// Expand input VALUE
+		val = GetFullPath(val)
+
+		//get total size of input files in MiB
+		info, err := os.Stat(val)
+		if os.IsNotExist(err) {
+			fmt.Fprintf(os.Stderr, "ERROR: Input file %s not found\n", val)
+		}
+		sizeMiB += (1.0 * float64(info.Size())) / (1024.0 * 1024.0) //fileinfo's Size() returns bytes, convert to MiB
+
+		// Replace key if found in args strings
+		// Handle replacing KEY or ${KEY} or $KEY
+		value := val
+		if directory, ok := tempDirectories[key]; ok {
+			value = directory  //replace with the temp directory if multiple files
+		}
+		seed.Job.Interface.Cmd = strings.Replace(seed.Job.Interface.Cmd,
+			"${"+key+"}", value, -1)
+		seed.Job.Interface.Cmd = strings.Replace(seed.Job.Interface.Cmd, "$"+key,
+			value, -1)
+		seed.Job.Interface.Cmd = strings.Replace(seed.Job.Interface.Cmd, key, value,
+			-1)
+
+		for _, k := range seed.Job.Interface.InputData.Files {
+			if k.Name == key {
+				if k.Multiple {
+					//directory has already been added to mount args, just link file into that directory
+					os.Link(val, filepath.Join(tempDirectories[key], info.Name()))
+				} else {
+					mountArgs = append(mountArgs, "-v")
+					mountArgs = append(mountArgs, val+":"+val)
+				}
+			}
+		}
+	}
+	
+	//remove unspecified unrequired inputs from cmd string
+	for _, k := range unrequired {
+		key := k
+		value := ""
+		seed.Job.Interface.Cmd = strings.Replace(seed.Job.Interface.Cmd,
+			"${"+key+"}", value, -1)
+		seed.Job.Interface.Cmd = strings.Replace(seed.Job.Interface.Cmd, "$"+key,
+			value, -1)
+		seed.Job.Interface.Cmd = strings.Replace(seed.Job.Interface.Cmd, key, value,
+			-1)
+	}
+
+	return mountArgs, sizeMiB, tempDirectories, nil
+}
+
+//SetOutputDir replaces the OUTPUT_DIR argument with the given output directory.
+// Returns output directory string
+func SetOutputDir(imageName string, seed *objects.Seed) string {
+
+	if !strings.Contains(seed.Job.Interface.Cmd, "OUTPUT_DIR") {
+		return ""
+	}
+
+	outputDir := runCmd.Lookup(constants.JobOutputDirFlag).Value.String()
+
+	// #37: if -o is not specified, and OUTPUT_DIR is in the command args,
+	//	auto create a time-stamped subdirectory with the name of the form:
+	//		imagename-iso8601timestamp
+	if outputDir == "" {
+		outputDir = "output-" + imageName + "-" + time.Now().Format(time.RFC3339)
+		outputDir = strings.Replace(outputDir, ":", "_", -1)
+	}
+
+	outdir := GetFullPath(outputDir)
+
+	// Check if outputDir exists. Create if not
+	if _, err := os.Stat(outdir); os.IsNotExist(err) {
+		// Create the directory
+		// Didn't find the specified directory
+		fmt.Fprintf(os.Stderr, "INFO: %s not found; creating directory...\n",
+			outdir)
+		os.Mkdir(outdir, os.ModePerm)
+	}
+
+	// Check if outdir is empty. Create time-stamped subdir if not
+	f, err := os.Open(outdir)
+	if err != nil {
+		// complain
+		fmt.Fprintf(os.Stderr, "ERROR: Error with %s. %s\n", outdir, err.Error())
+	}
+	defer f.Close()
+	_, err = f.Readdirnames(1)
+	if err != io.EOF {
+		// Directory is not empty
+		t := time.Now().Format("20060102_150405")
+		fmt.Fprintf(os.Stderr,
+			"INFO: Output directory %s is not empty. Creating sub-directory %s for Job Output Directory.\n",
+			outdir, t)
+		outdir = filepath.Join(outdir, t)
+		os.Mkdir(outdir, os.ModePerm)
+	}
+
+	seed.Job.Interface.Cmd = strings.Replace(seed.Job.Interface.Cmd,
+		"$OUTPUT_DIR", outdir, -1)
+	seed.Job.Interface.Cmd = strings.Replace(seed.Job.Interface.Cmd,
+		"${OUTPUT_DIR}", outdir, -1)
+	return outdir
+}
+
+//DefineMounts defines any seed specified mounts. TODO
+func DefineMounts(seed *objects.Seed) ([]string, error) {
+	// Ingest inputs into a map key = inputkey, value=inputpath
+	inputs := strings.Split(runCmd.Lookup(constants.MountFlag).Value.String(), ",")
+	inMap := make(map[string]string)
+	for _, f := range inputs {
+		x := strings.Split(f, "=")
+		if len(x) != 2 {
+			fmt.Fprintf(os.Stderr, "ERROR: Mount should be specified in KEY=VALUE format.\n")
+			fmt.Fprintf(os.Stderr, "ERROR: Unknown key for mount %v encountered.\n",
+				x)
+			continue
+		}
+		inMap[x[0]] = x[1]
+	}
+
+	// Valid by default
+	valid := true
+	var keys []string
+	for _, f := range seed.Job.Interface.Mounts {
+		keys = append(keys, f.Name)
+		if _, prs := inMap[f.Name]; !prs {
+			valid = false
+		}
+	}
+
+	if !valid {
+		var buffer bytes.Buffer
+		buffer.WriteString("ERROR: Incorrect mount key/values provided. -m arguments should be in the form:\n")
+		buffer.WriteString("  seed run -m MOUNT=path/to ...\n")
+		buffer.WriteString("The following mount keys are expected:\n")
+		for _, n := range keys {
+			buffer.WriteString("  " + n + "\n")
+		}
+		buffer.WriteString("\n")
+		return nil, errors.New(buffer.String())
+	}
+
+	var mounts []string
+	if seed.Job.Interface.Mounts != nil {
+		for _, mount := range seed.Job.Interface.Mounts {
+			mounts = append(mounts, "-v")
+			mountPath := inMap[mount.Name] + ":" + mount.Path
+
+			if mount.Mode != "" {
+				mountPath += ":" + mount.Mode
+			} else {
+				mountPath += ":ro"
+			}
+			mounts = append(mounts, mountPath)
+		}
+		return mounts, nil
+	}
+
+	return mounts, nil
+}
+
+//DefineSettings defines any seed specified docker settings. TODO
+// Return []string of docker command arguments in form of:
+//	"-?? setting1=val1 -?? setting2=val2 etc"
+func DefineSettings(seed *objects.Seed) ([]string, error) {
+	// Ingest inputs into a map key = inputkey, value=inputpath
+	inputs := strings.Split(runCmd.Lookup(constants.SettingFlag).Value.String(), ",")
+	inMap := make(map[string]string)
+	for _, f := range inputs {
+		x := strings.Split(f, "=")
+		if len(x) != 2 {
+			fmt.Fprintf(os.Stderr, "ERROR: Setting should be specified in KEY=VALUE format.\n")
+			fmt.Fprintf(os.Stderr, "ERROR: Unknown key for setting %v encountered.\n",
+				x)
+			continue
+		}
+		inMap[x[0]] = x[1]
+	}
+
+	// Valid by default
+	valid := true
+	var keys []string
+	for _, s := range seed.Job.Interface.Settings {
+		keys = append(keys, s.Name)
+		if _, prs := inMap[s.Name]; !prs {
+			valid = false
+		}
+
+	}
+
+	if !valid {
+		var buffer bytes.Buffer
+		buffer.WriteString("ERROR: Incorrect setting key/values provided. -e arguments should be in the form:\n")
+		buffer.WriteString("  seed run -e SETTING=somevalue ...\n")
+		buffer.WriteString("The following settings are expected:\n")
+		for _, n := range keys {
+			buffer.WriteString("  " + n + "\n")
+		}
+		buffer.WriteString("\n")
+		return nil, errors.New(buffer.String())
+	}
+
+	var settings []string
+	for _, key := range keys {
+		settings = append(settings, "-e")
+		settings = append(settings, GetNormalizedVariable(key)+"="+inMap[key])
+	}
+
+	return settings, nil
+}
+
+//DefineResources defines any seed specified docker resource requirements
+//based on the seed spec and the size of the input in MiB
+// returns array of arguments to pass to docker to restrict/specify the resources required
+// returns the total disk space requirement to be checked when validating output
+func DefineResources(seed *objects.Seed, inputSizeMiB float64) ([]string, float64, error) {
+	var resources []string
+	var disk float64
+
+	for _, s := range seed.Job.Interface.Resources.Scalar {
+		if s.Name == "mem" {
+			//resourceRequirement = inputVolume * inputMultiplier + constantValue
+			mem := (s.InputMultiplier * inputSizeMiB) + s.Value
+			intMem := int64(math.Ceil(mem)) //docker expects integer, get the ceiling of the specified value and convert
+			resources = append(resources, "-m")
+			resources = append(resources, fmt.Sprintf("%dm", intMem))
+		}
+		if s.Name == "disk" {
+			//resourceRequirement = inputVolume * inputMultiplier + constantValue
+			disk = (s.InputMultiplier * inputSizeMiB) + s.Value
+		}
+	}
+
+	return resources, disk, nil
+}
+
+//CheckRunOutput validates the output of the docker run command. Output data is
+// validated as defined in the seed.Job.Interface.OutputData.
+func CheckRunOutput(seed *objects.Seed, outDir string, diskLimit float64) {
+	// Validate any OutputData.Files
+	if seed.Job.Interface.OutputData.Files != nil {
+		fmt.Fprintf(os.Stderr, "INFO: Validating output files found under %s...\n",
+			outDir)
+
+		var dirSize int64
+		readSize := func(path string, file os.FileInfo, err error) error {
+			if !file.IsDir() {
+				dirSize += file.Size()
+			}
+
+			return nil
+		}
+		filepath.Walk(outDir, readSize)
+		sizeMB := float64(dirSize) / (1024.0 * 1024.0)
+		if diskLimit > 0 && sizeMB > diskLimit {
+			fmt.Fprintf(os.Stderr, "ERROR: Output directory exceeds disk space limit (%f MiB vs. %f MiB)\n", sizeMB, diskLimit)
+		}
+
+		// For each defined OutputData file:
+		//	#1 Check file media type
+		// 	#2 Check file names match output pattern
+		//  #3 Check number of files (if defined)
+		for _, f := range seed.Job.Interface.OutputData.Files {
+			// find all pattern matches in OUTPUT_DIR
+			matches, _ := filepath.Glob(path.Join(outDir, f.Pattern))
+
+			// Check media type of matches
+			count := 0
+			var matchList []string
+			for _, match := range matches {
+				ext := filepath.Ext(match)
+				mType := mime.TypeByExtension(ext)
+				if strings.Contains(mType, f.MediaType) ||
+					strings.Contains(f.MediaType, mType) {
+					count++
+					matchList = append(matchList, "\t"+match+"\n")
+					metadata := match + ".metadata.json"
+					if _, err := os.Stat(metadata); err == nil {
+						err := ValidateSeedFile("", metadata, constants.SchemaMetadata)
+						if err != nil {
+							fmt.Fprintf(os.Stderr, "ERROR: Side-car metadata file %s validation error: %s", metadata, err.Error())
+						}
+					}
+				}
+			}
+
+			// Validate number of matches to specified number
+			if f.Count != "" && f.Count != "*" && f.Required {
+				count, _ := strconv.Atoi(f.Count)
+				if count != len(matchList) {
+					fmt.Fprintf(os.Stderr, "ERROR: %v files specified, %v found.\n",
+						f.Count, strconv.Itoa(len(matchList)))
+					if len(matchList) > 0 {
+						for _, s := range matchList {
+							fmt.Fprintf(os.Stderr, s)
+						}
+					}
+				} else {
+					fmt.Fprintf(os.Stderr, "SUCCESS: %v files specified, %v found. Files found:\n",
+						f.Count, strconv.Itoa(len(matchList)))
+					for _, s := range matchList {
+						fmt.Fprintf(os.Stderr, s)
+					}
+				}
+			}
+		}
+	}
+
+	// Validate any defined OutputData.Json
+	// Look for ResultsFileManifestName.json in the root of the OUTPUT_DIR
+	// and then validate any keys identified in OutputData exist
+	if seed.Job.Interface.OutputData.JSON != nil {
+		fmt.Fprintf(os.Stderr, "INFO: Validating %s...\n",
+			filepath.Join(outDir, constants.ResultsFileManifestName))
+		// look for results manifest
+		manfile := filepath.Join(outDir, constants.ResultsFileManifestName)
+		if _, err := os.Stat(manfile); os.IsNotExist(err) {
+			fmt.Fprintf(os.Stderr, "ERROR: %s specified but cannot be found. %s\n Exiting testrunner.\n",
+				constants.ResultsFileManifestName, err.Error())
+			return
+		}
+
+		bites, err := ioutil.ReadFile(filepath.Join(outDir,
+			constants.ResultsFileManifestName))
+		if err != nil {
+			fmt.Fprintf(os.Stderr, "ERROR: Error reading %s.%s\n",
+				constants.ResultsFileManifestName, err.Error())
+			return
+		}
+
+		documentLoader := gojsonschema.NewStringLoader(string(bites))
+		_, err = documentLoader.LoadJSON()
+		if err != nil {
+			fmt.Fprintf(os.Stderr, "ERROR: Error loading results manifest file: %s. %s\n Exiting testrunner.\n",
+				constants.ResultsFileManifestName, err.Error())
+			return
+		}
+
+		schemaFmt := "{ \"type\": \"object\", \"properties\": { %s }, \"required\": [ %s ] }"
+		schema := ""
+		required := ""
+
+		// Loop through defined name/key values to extract from results_manifest.json
+		for _, jsonStr := range seed.Job.Interface.OutputData.JSON {
+			key := jsonStr.Name
+			if jsonStr.Key != "" {
+				key = jsonStr.Key
+			}
+
+			schema += fmt.Sprintf("\"%s\": { \"type\": \"%s\" },", key, jsonStr.Type)
+
+			if jsonStr.Required {
+				required += fmt.Sprintf("\"%s\",", key)
+			}
+		}
+		//remove trailing commas
+		if len(schema) > 0 {
+			schema = schema[:len(schema)-1]
+		}
+		if len(required) > 0 {
+			required = required[:len(required)-1]
+		}
+
+		schema = fmt.Sprintf(schemaFmt, schema, required)
+
+		schemaLoader := gojsonschema.NewStringLoader(schema)
+		schemaResult, err := gojsonschema.Validate(schemaLoader, documentLoader)
+		if err != nil {
+			fmt.Fprintf(os.Stderr, "ERROR: Error running validator: %s\n Exiting testrunner.\n",
+				err.Error())
+			return
+		}
+
+		if len(schemaResult.Errors()) == 0 {
+			fmt.Fprintf(os.Stderr, "SUCCESS: Results manifest file is valid.\n")
+		}
+
+		for _, desc := range schemaResult.Errors() {
+			fmt.Fprintf(os.Stderr, "ERROR: %s is invalid: - %s\n", constants.ResultsFileManifestName, desc)
+		}
+	}
+}
+
+//DefineRunFlags defines the flags for the seed run command
+func DefineRunFlags() {
+	runCmd = flag.NewFlagSet(constants.RunCommand, flag.ContinueOnError)
+
+	var imgNameFlag string
+	runCmd.StringVar(&imgNameFlag, constants.ImgNameFlag, "",
+		"Name of Docker image to run")
+	runCmd.StringVar(&imgNameFlag, constants.ShortImgNameFlag, "",
+		"Name of Docker image to run")
+
+	var settings objects.ArrayFlags
+	runCmd.Var(&settings, constants.SettingFlag,
+		"Defines the value to be applied to setting")
+	runCmd.Var(&settings, constants.ShortSettingFlag,
+		"Defines the value to be applied to setting")
+
+	var mounts objects.ArrayFlags
+	runCmd.Var(&mounts, constants.MountFlag,
+		"Defines the full path to be mapped via mount")
+	runCmd.Var(&mounts, constants.ShortMountFlag,
+		"Defines the full path to be mapped via mount")
+
+	var inputs objects.ArrayFlags
+	runCmd.Var(&inputs, constants.InputDataFlag,
+		"Defines the full path to any input data arguments")
+	runCmd.Var(&inputs, constants.ShortInputDataFlag,
+		"Defines the full path to input data arguments")
+
+	var outdir string
+	runCmd.StringVar(&outdir, constants.JobOutputDirFlag, "",
+		"Full path to the algorithm output directory")
+	runCmd.StringVar(&outdir, constants.ShortJobOutputDirFlag, "",
+		"Full path to the algorithm output directory")
+
+	var metadataSchema string
+	runCmd.StringVar(&metadataSchema, constants.SchemaFlag, "",
+		"Metadata schema file to override built in schema in validating side-car metadata files")
+	runCmd.StringVar(&metadataSchema, constants.ShortSchemaFlag, "",
+		"Metadata schema file to override built in schema in validating side-car metadata files")
+
+	var rmVar bool
+	runCmd.BoolVar(&rmVar, constants.RmFlag, false,
+		"Specifying the -rm flag automatically removes the image after executing docker run")
+
+	// Run usage function
+	runCmd.Usage = func() {
+		PrintRunUsage()
+	}
+}
+
+//PrintRunUsage prints the seed run usage arguments, then exits the program
+func PrintRunUsage() {
+	fmt.Fprintf(os.Stderr,
+		"\nUsage:\tseed run -in IMAGE_NAME [-i INPUT_KEY=INPUT_FILE ...] [-e SETTING_KEY=VALUE] -o OUTPUT_DIRECTORY \n")
+
+	fmt.Fprintf(os.Stderr, "\nRuns Docker image defined by seed spec.\n")
+
+	fmt.Fprintf(os.Stderr, "\nOptions:\n")
+	fmt.Fprintf(os.Stderr, "  -%s  -%s \t Specifies the key/value setting values of the seed spec in the format SETTING_KEY=VALUE\n",
+		constants.ShortSettingFlag, constants.SettingFlag)
+	fmt.Fprintf(os.Stderr, "  -%s  -%s Specifies the key/value input data values of the seed spec in the format INPUT_FILE_KEY=INPUT_FILE_VALUE\n",
+		constants.ShortInputDataFlag, constants.InputDataFlag)
+	fmt.Fprintf(os.Stderr, "  -%s -%s Docker image name to run\n",
+		constants.ShortImgNameFlag, constants.ImgNameFlag)
+	fmt.Fprintf(os.Stderr, "  -%s  -%s \t Specifies the key/value mount values of the seed spec in the format MOUNT_KEY=HOST_PATH\n",
+		constants.ShortMountFlag, constants.MountFlag)
+	fmt.Fprintf(os.Stderr, "  -%s  -%s \t Job Output Directory Location\n",
+		constants.ShortJobOutputDirFlag, constants.JobOutputDirFlag)
+	fmt.Fprintf(os.Stderr, "  -%s \t\t Automatically remove the container when it exits (docker run --rm)\n",
+		constants.RmFlag)
+	fmt.Fprintf(os.Stderr, "  -%s  -%s \t External Seed metadata schema file; Overrides built in schema to validate side-car metadata files\n",
+		constants.ShortSchemaFlag, constants.SchemaFlag)
+	os.Exit(0)
+}
